requirements.txt
flask
pymysql
flask-mysql
argon2-cffi